Move all keyboard relative code to the keyboard interface
