+++ yabause/src/perlinuxjoy.c
@@ -33,6 +33,8 @@ u32  PERLinuxJoyScan(u32 flags);
 void PERLinuxJoyFlush(void);
 void PERLinuxKeyName(u32 key, char * name, int size);
 
+int getSupportedJoy(const char *name);
+
 #define SERVICE_BUTTON_NUMBER 3
 #define SERVICE_BUTTON_EXIT (PERGUN_AXIS+1)
 #define SERVICE_BUTTON_TOGGLE (PERGUN_AXIS+2)
@@ -156,6 +158,46 @@ joymapping_struct joyMapping[MAPPING_NB] = {
    },
 };
 
+#define KEYPAD(key, player) ((player << 17)|key)
+
+static void KeyInit() {
+  void * padbits;
+
+  PerPortReset();
+  padbits = PerPadAdd(&PORTDATA1);
+
+  PerSetKey(KEYPAD(PERPAD_UP, 0), PERPAD_UP, padbits);
+  PerSetKey(KEYPAD(PERPAD_RIGHT, 0), PERPAD_RIGHT, padbits);
+  PerSetKey(KEYPAD(PERPAD_DOWN, 0), PERPAD_DOWN, padbits);
+  PerSetKey(KEYPAD(PERPAD_LEFT, 0), PERPAD_LEFT, padbits);
+  PerSetKey(KEYPAD(PERPAD_RIGHT_TRIGGER, 0), PERPAD_RIGHT_TRIGGER, padbits);
+  PerSetKey(KEYPAD(PERPAD_LEFT_TRIGGER, 0), PERPAD_LEFT_TRIGGER, padbits);
+  PerSetKey(KEYPAD(PERPAD_START, 0), PERPAD_START, padbits);
+  PerSetKey(KEYPAD(PERPAD_A, 0), PERPAD_A, padbits);
+  PerSetKey(KEYPAD(PERPAD_B, 0), PERPAD_B, padbits);
+  PerSetKey(KEYPAD(PERPAD_C, 0), PERPAD_C, padbits);
+  PerSetKey(KEYPAD(PERPAD_X, 0), PERPAD_X, padbits);
+  PerSetKey(KEYPAD(PERPAD_Y, 0), PERPAD_Y, padbits);
+  PerSetKey(KEYPAD(PERPAD_Z, 0), PERPAD_Z, padbits);
+
+  padbits = PerPadAdd(&PORTDATA2);
+
+  PerSetKey(KEYPAD(PERPAD_UP, 1), PERPAD_UP, padbits);
+  PerSetKey(KEYPAD(PERPAD_RIGHT, 1), PERPAD_RIGHT, padbits);
+  PerSetKey(KEYPAD(PERPAD_DOWN, 1), PERPAD_DOWN, padbits);
+  PerSetKey(KEYPAD(PERPAD_LEFT, 1), PERPAD_LEFT, padbits);
+  PerSetKey(KEYPAD(PERPAD_RIGHT_TRIGGER, 1), PERPAD_RIGHT_TRIGGER, padbits);
+  PerSetKey(KEYPAD(PERPAD_LEFT_TRIGGER, 1), PERPAD_LEFT_TRIGGER, padbits);
+  PerSetKey(KEYPAD(PERPAD_START, 1), PERPAD_START, padbits);
+  PerSetKey(KEYPAD(PERPAD_A, 1), PERPAD_A, padbits);
+  PerSetKey(KEYPAD(PERPAD_B, 1), PERPAD_B, padbits);
+  PerSetKey(KEYPAD(PERPAD_C, 1), PERPAD_C, padbits);
+  PerSetKey(KEYPAD(PERPAD_X, 1), PERPAD_X, padbits);
+  PerSetKey(KEYPAD(PERPAD_Y, 1), PERPAD_Y, padbits);
+  PerSetKey(KEYPAD(PERPAD_Z, 1), PERPAD_Z, padbits);
+
+}
+
 int requestExit = 0;
 int toggle = 0;
 int service = 0;
@@ -248,6 +290,7 @@ static int LinuxJoyInit(perlinuxjoy_struct * joystick, const char * path, int id
    char name[128];
    struct js_event evt;
    ssize_t num_read;
+
    joystick->fd = open(path, O_RDONLY | O_NONBLOCK);
    if (joystick->fd == -1) return -1;
    if (ioctl(joystick->fd, JSIOCGNAME(sizeof(name)), name) < 0)
@@ -273,11 +316,14 @@ static int LinuxJoyInit(perlinuxjoy_struct * joystick, const char * path, int id
    }
 
    if (joystick->axiscount > MAXAXIS) joystick->axiscount = MAXAXIS;
+
    joystick->axis = malloc(sizeof(int) * joystick->axiscount);
    for(i = 0;i < joystick->axiscount;i++)
    {
       joystick->axis[i] = axisinit[i];
    }
+
+   KeyInit();
    return 0;
 }
 
@@ -293,6 +339,7 @@ static void LinuxJoyHandleEvents(perlinuxjoy_struct * joystick)
 {
    struct js_event evt;
    ssize_t num_read;
+   int key;
 
    if (joystick->fd == -1) return;
 
@@ -312,7 +359,7 @@ static void LinuxJoyHandleEvents(perlinuxjoy_struct * joystick)
          else if (evt.value < initvalue) evt.value = -1;
          else evt.value = 1;
       }
-      int key = PACKEVENT(evt, joystick);
+      key = PACKEVENT(evt, joystick);
       if (evt.value != 0)
       {
          if ((key & 0x1FFFF) != 0x1FFFF) PerKeyDown(key);
@@ -329,6 +376,7 @@ static int LinuxJoyScan(perlinuxjoy_struct * joystick)
 {
    struct js_event evt;
    ssize_t num_read;
+   int key;
 
    if (joystick->fd == -1) return 0;
 
@@ -348,7 +396,7 @@ static int LinuxJoyScan(perlinuxjoy_struct * joystick)
       else if (evt.value < initvalue) evt.value = -1;
       else evt.value = 1;
    }
-   int key = PACKEVENT(evt, joystick);
+   key = PACKEVENT(evt, joystick);
    if ((key & 0x1FFFF) != 0x1FFFF)
       return key;
    else return 0;
@@ -379,7 +427,9 @@ int PERLinuxJoyInit(void)
    int fd;
    int j=0;
    glob_t globbuf;
+
    glob("/dev/input/js*", 0, NULL, &globbuf);
+
    joycount = globbuf.gl_pathc;
    joysticks = calloc(sizeof(perlinuxjoy_struct), joycount);
    for(i = 0;i < globbuf.gl_pathc;i++) {
@@ -388,6 +438,7 @@ int PERLinuxJoyInit(void)
    }
    joycount = j;
    globfree(&globbuf);
+
    return 0;
 }
 
@@ -408,6 +459,7 @@ void PERLinuxJoyDeInit(void)
 int PERLinuxJoyHandleEvents(void)
 {
    int i;
+
    for(i = 0;i < joycount;i++)
       LinuxJoyHandleEvents(joysticks + i);
